requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.1
sqlite3
python-dotenv==1.0.0
